WIP Day 4
Added possibility to register in an arena with only pressing a button and it does all actions to register, request tokens etc., a move pad cheat so you can play even without a bot and move (and attack) with buttons. Added a marker to localize your bot in the grid. Added joystick to move camera from mobile phones and some minor fixes and improvements